// src/model/dto/RendaFixaDTO.tsx
interface RendaFixaDTO {
    nome: string;
    precoAtual: number;
    precoCompra: number;
    dataCadastro: string;
    precoMinimo: number;
    precoMaximo: number;
    codigo: string;
    taxaRetorno: number;
    dataVencimento: string;
    precoVenda: number;
    usuarioEmail: string;
}

export default RendaFixaDTO