// src/model/dto/FundoImobiliarioDTO.tsx
interface FundoImobiliarioDTO {
    nome: string;
    precoAtual: number;
    precoCompra: number;
    dataCadastro: string;
    precoMinimo: number;
    precoMaximo: number;
    codigoFii: string;
    rendimentoMensal: number;
    precoVenda: number;
    usuarioEmail: string;
}

export default FundoImobiliarioDTO